+++ src/health/rrdcalc.c
@@ -148,7 +148,7 @@ static STRING *rrdcalc_replace_variables_with_rrdset_labels(const char *line, RR
 
 void rrdcalc_update_info_using_rrdset_labels(RRDCALC *rc) {
     if(rc->rrdset && rc->rrdset->rrdlabels) {
-        size_t labels_version = rrdlabels_version(rc->rrdset->rrdlabels);
+        uint32_t labels_version = rrdlabels_version(rc->rrdset->rrdlabels);
         if (rc->labels_version != labels_version) {
             STRING *old;
